requirements.txt
fire==0.1.2
future==0.16.0
osaapi==0.3.12
requests==2.20.1
PyYAML==3.13
six==1.11.0
backports.tempfile==1.0
